pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "commit0"
version = "0.1.8"
description = "A development and evaluation framework for using language models to generate libraries."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.6.4",
    "pre-commit>=3.8.0",
    "typer>=0.12.0",
    "docker>=7.1.0",
    "fastcore>=1.7.8",
    "ghapi>=1.0.6",
    "gitpython>=3.1.43",
    "pytest>=8.3.3",
    "datasets>=3.0.1",
    "modal>=0.66.26",
    "strenum>=0.4.15",
    "e2b-code-interpreter>=1.0.4",
    "python-dotenv>=1.0.1",
    "boto3>=1.34.0",
]
classifiers = [
    "License :: OSI Approved :: MIT License",
]

[project.scripts]
commit0 = "commit0.__main__:main"
agent = "agent.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["commit0", "agent"]

[project.optional-dependencies]
agent = [
    "import-deps>=0.3.0",
    "PyMuPDF>=1.24.5",
    "aider-chat",
]
scrape = [
    "playwright>=1.40.0",
    "PyMuPDF>=1.24.5",
    "PyPDF2>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://commit-0.github.io/"
Repository = "https://github.com/commit-0/commit0"
Issues = "https://github.com/commit-0/commit0/issues"

[tool.pyright]
include = ["**/commit0", "**/agent", "**/tools"]
ignore = ["**/docs", "**/logs"]
venvPath = "."
venv = ".venv"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownArgumentType = "none"
reportUnknownVariableType = "none"
reportMissingTypeArgument = "none"
reportMissingTypeStubs = "none"
reportUnusedExpression = "none"
reportUnknownLambdaType = "none"
reportIncompatibleMethodOverride = "none"
reportPrivateUsage = "none"
reportMissingParameterType = "error"


[tool.ruff]
exclude = [
    ".git",
    "**/.venv/**",
    "__pycache__",
    "**/docs/**/*",
    "**/logs/**/*",
]

lint.ignore = [
    "ANN101",
    "ANN401",
    "N801",
    "E203",
    "E266",
    "E501",
    "E741",
    "N803",
    "N802",
    "N806",
    "D400",
    "D401",
    "D105",
    "D415",
    "D402",
    "D205",
    "D100",
    "D101",
    "D107",
    "D203",
    "D213",
    "ANN204",
    "ANN102",
]
lint.select = ["D", "E", "F", "N", "ANN"]
lint.fixable = [
    "A",
    "B",
    "C",
    "D",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "T",
    "W",
    "ANN",
    "ARG",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "EM",
    "ERA",
    "EXE",
    "FBT",
    "ICN",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
lint.unfixable = []

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.uv]
override-dependencies = [
    "litellm",
]

[tool.uv.sources]
aider-chat = { git = "https://github.com/Ethara-Ai/aider.git", branch = "main" }
litellm = { git = "https://github.com/Ethara-Ai/litellm.git", branch = "main" }
